refactor: Model 디렉토리에 커스텀 클래스 파일 생성
Issue: #23
