+++ ios/SideDishApp/SideDishApp/Model/Category.swift
@@ -1,0 +1,15 @@
+//
+//  Category.swift
+//  SideDishApp
+//
+//  Created by Cory Kim on 2020/04/24.
+//  Copyright © 2020 corykim0829. All rights reserved.
+//
+
+import Foundation
+
+struct Category: Decodable {
+    let name: String
+    let description: String
+    let path: String
+}
