+++ ios/SideDishApp/SideDishApp/Network/NetworkManager.swift
@@ -28,47 +28,3 @@ class NetworkManager {
         }
     }
 }
-
-struct Category: Decodable {
-    let name: String
-    let description: String
-    let path: String
-}
-
-struct ProductList: Decodable {
-    let products: [Product]
-    
-    enum CodingKeys: String, CodingKey {
-        case products = "body"
-    }
-}
-
-struct Products {
-    let products: [Product]
-    
-    var count: Int {
-        get { products.count }
-    }
-    
-    init(_ products: [Product] = []) {
-        self.products = products
-    }
-}
-
-struct Product: Decodable {
-    let detailHash: String
-    let imageURL: String
-    let title: String
-    let originalPrice: String?
-    let finalPrice: String
-    let badges: [String]?
-    
-    enum CodingKeys: String, CodingKey {
-        case detailHash = "detail_hash"
-        case imageURL = "image"
-        case title
-        case originalPrice = "n_price"
-        case finalPrice = "s_price"
-        case badges = "badge"
-    }
-}
